refactor(hmsl): move input and output elements to dedicated files
