+++ ggshield/cmd/hmsl/fingerprint.py
@@ -1,7 +1,6 @@
-from typing import Any, Dict, Iterable, Iterator, Set, TextIO, cast
+from typing import Any, TextIO, cast
 
 import click
-from dotenv import dotenv_values
 
 from ggshield.cmd.common_options import add_common_options
 from ggshield.cmd.hmsl.hmsl_common_options import (
@@ -13,10 +12,8 @@ from ggshield.cmd.hmsl.hmsl_common_options import (
     naming_strategy_option,
 )
 from ggshield.core.text_utils import display_info
-from ggshield.hmsl.client import PREFIX_LENGTH
-from ggshield.hmsl.collection import PreparedSecrets, SecretWithKey
-from ggshield.hmsl.crypto import hash_string
-from ggshield.hmsl.utils import EXCLUDED_KEYS, EXCLUDED_VALUES
+from ggshield.hmsl.collection import collect, prepare
+from ggshield.hmsl.output import write_outputs
 
 
 def validate_prefix(prefix: str) -> str:
@@ -59,71 +56,3 @@ def fingerprint_cmd(
 
     display_info(f"Prepared {len(secrets)} secrets.")
     return 0
-
-
-# Helper methods
-
-
-def collect(
-    input: TextIO, input_type: InputType = InputType.FILE
-) -> Iterator[SecretWithKey]:
-    """
-    Collect the secrets
-    """
-    if input_type == InputType.ENV:
-        config = dotenv_values(stream=input)
-        for key, value in config.items():
-            # filter our excluded keys and values
-            if not key or not value:
-                continue
-            if key.upper() in EXCLUDED_KEYS or value.lower() in EXCLUDED_VALUES:
-                continue
-            yield SecretWithKey(value=value, key=key)
-    else:
-        for line in input:
-            secret = line.strip()
-            if secret == "":
-                # Skip empty lines
-                continue
-            yield SecretWithKey(value=secret, key=None)
-
-
-def prepare(
-    secrets: Iterable[SecretWithKey],
-    naming_strategy: NamingStrategy,
-    *,
-    full_hashes: bool = False,
-) -> PreparedSecrets:
-    """
-    Prepare the secrets so they can later be checked.
-    """
-    hashes: Set[str] = set()
-    mapping: Dict[str, str] = {}
-    for secret in secrets:
-        name = naming_strategy(secret)
-        hash = hash_string(secret.value)
-        mapping[hash] = name
-        if full_hashes:
-            hashes.add(hash)
-        else:
-            hashes.add(hash[:PREFIX_LENGTH])
-    return PreparedSecrets(
-        payload=hashes,
-        mapping=mapping,
-    )
-
-
-def write_outputs(result: PreparedSecrets, prefix: str) -> None:
-    """
-    Write payload and mapping files.
-    """
-    with open(f"{prefix}payload.txt", "w") as payload_file:
-        payload_file.write("\n".join(result.payload) + "\n")
-
-    with open(f"{prefix}mapping.txt", "w") as mapping_file:
-        for hash, hint in result.mapping.items():
-            line = hash + ":" + hint if hint else hash
-            mapping_file.write(line + "\n")
-    display_info(
-        f"{prefix}payload.txt and {prefix}mapping.txt files have been written."
-    )
