+++ ggshield/hmsl/collection.py
@@ -1,5 +1,12 @@
 from dataclasses import dataclass
-from typing import Dict, Optional, Set
+from typing import Dict, Iterable, Iterator, Optional, Set, TextIO
+
+from dotenv import dotenv_values
+
+from ggshield.cmd.hmsl.hmsl_common_options import InputType, NamingStrategy
+from ggshield.hmsl import PREFIX_LENGTH
+from ggshield.hmsl.crypto import hash_string
+from ggshield.hmsl.utils import EXCLUDED_KEYS, EXCLUDED_VALUES
 
 
 @dataclass
@@ -12,3 +19,52 @@ class PreparedSecrets:
 class SecretWithKey:
     key: Optional[str]
     value: str
+
+
+def collect(
+    input: TextIO, input_type: InputType = InputType.FILE
+) -> Iterator[SecretWithKey]:
+    """
+    Collect the secrets
+    """
+    if input_type == InputType.ENV:
+        config = dotenv_values(stream=input)
+        for key, value in config.items():
+            # filter our excluded keys and values
+            if not key or not value:
+                continue
+            if key.upper() in EXCLUDED_KEYS or value.lower() in EXCLUDED_VALUES:
+                continue
+            yield SecretWithKey(value=value, key=key)
+    else:
+        for line in input:
+            secret = line.strip()
+            if secret == "":
+                # Skip empty lines
+                continue
+            yield SecretWithKey(value=secret, key=None)
+
+
+def prepare(
+    secrets: Iterable[SecretWithKey],
+    naming_strategy: NamingStrategy,
+    *,
+    full_hashes: bool = False,
+) -> PreparedSecrets:
+    """
+    Prepare the secrets so they can later be checked.
+    """
+    hashes: Set[str] = set()
+    mapping: Dict[str, str] = {}
+    for secret in secrets:
+        name = naming_strategy(secret)
+        hash = hash_string(secret.value)
+        mapping[hash] = name
+        if full_hashes:
+            hashes.add(hash)
+        else:
+            hashes.add(hash[:PREFIX_LENGTH])
+    return PreparedSecrets(
+        payload=hashes,
+        mapping=mapping,
+    )
